Fix warnings in bits.hh

+++ src/clean-core/bits.hh
@@ -11,40 +11,40 @@ namespace cc
 {
 #ifdef CC_COMPILER_MSVC
 
-inline int popcount(uint8 v) { return __popcnt16(v); }
-inline int popcount(uint16 v) { return __popcnt16(v); }
-inline int popcount(uint32 v) { return __popcnt(v); }
-inline int popcount(uint64 v) { return __popcnt64(v); }
+inline int popcount(uint8 v) { return static_cast<int>(__popcnt16(v)); }
+inline int popcount(uint16 v) { return static_cast<int>(__popcnt16(v)); }
+inline int popcount(uint32 v) { return static_cast<int>(__popcnt(v)); }
+inline int popcount(uint64 v) { return static_cast<int>(__popcnt64(v)); }
 
 inline int count_trailing_zeros(uint8 v)
 {
     unsigned long idx;
     _BitScanForward(&idx, v);
-    return idx;
+    return static_cast<int>(idx);
 }
 inline int count_trailing_zeros(uint16 v)
 {
     unsigned long idx;
     _BitScanForward(&idx, v);
-    return idx;
+    return static_cast<int>(idx);
 }
 inline int count_trailing_zeros(uint32 v)
 {
     unsigned long idx;
     _BitScanForward(&idx, v);
-    return idx;
+    return static_cast<int>(idx);
 }
 inline int count_trailing_zeros(uint64 v)
 {
     unsigned long idx;
     _BitScanForward64(&idx, v);
-    return idx;
+    return static_cast<int>(idx);
 }
 
-inline int count_leading_zeros(uint8 v) { return __lzcnt16(v) - 8; }
-inline int count_leading_zeros(uint16 v) { return __lzcnt16(v); }
-inline int count_leading_zeros(uint32 v) { return __lzcnt(v); }
-inline int count_leading_zeros(uint64 v) { return __lzcnt64(v); }
+inline int count_leading_zeros(uint8 v) { return static_cast<int>(__lzcnt16(v) - 8); }
+inline int count_leading_zeros(uint16 v) { return static_cast<int>(__lzcnt16(v)); }
+inline int count_leading_zeros(uint32 v) { return static_cast<int>(__lzcnt(v)); }
+inline int count_leading_zeros(uint64 v) { return static_cast<int>(__lzcnt64(v)); }
 
 #else
